refactor(crypto): move ByteBuffer under utils
