+++ packages/crypto/src/transactions/serializer.ts
@@ -1,4 +1,3 @@
-import { ByteBuffer } from "../byte-buffer";
 import { TransactionType, TransactionTypeGroup } from "../enums";
 import { TransactionVersionError } from "../errors";
 import { Address } from "../identities";
@@ -6,7 +5,7 @@ import { ISerializeOptions } from "../interfaces";
 import { ITransaction, ITransactionData } from "../interfaces";
 import { configManager } from "../managers/config";
 import { isException } from "../utils";
-import { isSupportedTransactionVersion } from "../utils";
+import { ByteBuffer, isSupportedTransactionVersion } from "../utils";
 import { TransactionTypeFactory } from "./types";
 
 // Reference: https://github.com/ArkEcosystem/AIPs/blob/master/AIPS/aip-11.md
