+++ packages/crypto/src/utils/byte-buffer.ts
